Example created by ChatGPT
Here is example code made by ChatGPT creating a neural network to classify text data as negative, neutral, or positive. Could be a useful starting point. I think we might want to change the y variable to be a length 3 vector like [0, 1, 0] rather than just a number 0, 1, or 2. But not sure at this point.